patch the logic to generate a filesystem path for the requesting user
The previous logic only worked if the authenticated client id
contains a `/` so that it can be split into two parts.

The local database id provider would not normally have such
id values, so revise the logic to tolerate this.

This is still a bit fragile, but allows some testing without
a full globus integration.